pkg: utop dev tool
This adds utop as a dev tool. Unlike most dev tools there is no "dune
tools exec utop" command, but rather the existing "dune utop" command
has been modified such that if a lockdir exists for the utop dev tool,
then dune will install the utop library (downloading it if necessary)
within the _build directory and use that library to create the custom
utop toplevel for the project.

It's possible to install utop as a dev tool without running it by
running the "dune tools install utop" command.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>