pkg: command to check if project uses package management (#11997)
Some users have been using the presence of a lockdir to determine
whether a project indicates that package management should be used, for
example when a text editor checks whether ocamllsp should be invoked via
dune (as a dev tool) or from an opam switch. This is brittle, as dune
may change the way projects indicate they intend to use package
management in the future.

This change introduces a new command "dune pkg enabled" which
communicates via its exit status whether dune detects that a project
intends to use package management (a status of 0 indicates package
management is enabled). If/when dune changes how a project should
indicate that it should use package management, the implementation of
this command will be changed to match.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>