flake: add nix targets that implement `dune --version`
Existing nix targets produce binaries whose version according to `dune
--version` is "n/a". This adds a mechanism to the flake for determining
the version number and building a dune binary which outputs the
appropriate version number from `dune --version`.

Note that this depends on two hacks which are explained in more detail
in a comment:
 - determining the version number from the changelog file
 - creating a temporary git repo with a tag named after the version to
   allow `dune subst` to find out the current version

In the longer term we could remove the needs for these workarounds. In
the short term this feature is needed to determine the current version
of dune when dune is shipped as a binary, as we use the flake to build
said binaries.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>