pkg: lock utop dev tool before setting up utop rules
This reverts commit 6c8a5a339c7cdbcda6cc70ba0556fc5eec765977 which
locked utop in a separate fiber scheduler run. Instead we just make sure
that the lockdir is created (if necessary) before loading the utop
rules, which happens as a side effect of checking whether the utop exe
exists, and _has_ the side effect of checking and memoizing whether the
utop lockdir exists.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>