feat(exec): support absolute paths
We add support for absolute paths passed to dune exec. There is quite
some cleanup that needs to be done in the path handling in dune exec,
but for now we have kept this change small and tried to record some of
its behaviour.

A current limitation is that if a watch mode dune is running then dune
exec will connect via RPC and will not be able to correctly deal with an
absolute path.

Signed-off-by: Ali Caglayan <alizter@gmail.com>