Replace `git` function by environment variables
As #12104 shows, it turns out that exporting variables when a function
is called is not POSIX. Thus calls of the form `FOO=bar git` are not
portable between systems.

Fortunately, git supports variables to trigger overriding the
author/committer, thus instead of creating a wrapper function we're
setting these variables.

Signed-off-by: Marek Kubica <marek@tarides.com>