pkg: don't set %{jobs} variable from -j in deps
When building a project's dependencies, setting the value of %{jobs}
from the value of -j means that each time a different value is passed to
${jobs}, each dependency of the project that references the %{jobs}
variable in its dependency cone is rebuilt. This includes the compiler
which can take several minutes to build. Setting the %{jobs} variable to
a fixed value (chosen as the default degree of concurrency used by dune)
avoids this issue. Passing a value to -j still affects the degree of
concurrency used when building the project itself (as opposed to its
dependencies).

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>