feat(tui): wrap long messages
We leverage the underlying format mechanism to allow for word wrapping.
Previously, a user would have to rely on the scroll bars in order to
view wide text. For most breakable text we can do better by breaking up
the text with Format. Verbatim strings that are longer than the width
will still trigger the horizontal scroll bar as usual.

Signed-off-by: Ali Caglayan <alizter@gmail.com>