refactor: move display to [Action_intf.Intf.Exec]
This is both less code, and is closer to what the fork is doing.
Moreover, maintaining an extra field in a record is less work to sync
than a variable that gets passed to just about every other function.

As a reminder, some actions need to observe this variable to respect
[--display foo]. It is not possible to do this inside [Process] itself
because some process execution functions do not respect this flag (on
purppose)

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>

<!-- ps-id: 9442ddeb-267d-4091-9f33-e508fb1d23fc -->