Change lock dir path type to `Path.t`
Given we plan to move the lock dirs into the build directory, we can't
continue to use `Path.Source.t`. We can either move them to
`Path.Build.t` or `Path.t`.

The former has the advantage that it restricts the system from writing
the lock dir to any place that isn't the build dir, however it requires
a lot more changes that can't meaningfully be unbundled from #11775.

This PR is more conservative and only moves it to `Path.t`. It still
relies on the fact that the `Path.t` is not an `Path.Build.t` but most
of them are backed by a `Path.Source.t`. It also adds a bit of
lazy-loading as `Path.of_string` requires `Fgetcl` to be initialized so
the parsing can't happen at CLI parse time, but the complication is
quite marginal.

Signed-off-by: Marek Kubica <marek@tarides.com>