feature: allow custom actions to ask for more dependencies (#12189)
The goal of this change is to extract minimal interface of the dune
action plugin so that its current implementation can be moved out of the
engine.

To do this, we just allow custom actions to return the
[Done_or_more_deps.t] type just like every other "built-in" action can.
There was no good reason for this limitation anyway. The only reason for
the inconsistency was that custom actions were added without much
thought for DAP.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>