pkg: improve performance of null builds (#12248)
Between now and the investigation of this slowness described in the issue, I have
forgotten what's happening exactly, but regardless this Pkg_rules.DB.t type is large,
which makes comparing it with itself costly. By avoiding recomputing the same
datastructure multiple times, we can cutoff the comparison with physical equality.

`dune build -j 1` in the test repo from the issue goes from 4s to 2s or so.

Signed-off-by: Valentin Gatien-Baron <valentin.gatienbaron@gmail.com>