Avoid using Poly.hash on Set and Map in pkg_rules
If two values are equal then their hashes must also be equal.
`Hashtbl.hash` (or `Poly.hash` as its known in dune) can compute a hash
for an arbitrary value by traversing its representation in memory, the
so-called "structure" of the value. Thus if two values differ in their
structure, their hashes will likely also differ.

Some OCaml types have a notion of equality which is looser than
structural equality. That is, two values may be considered equal even if
their representations differ. Common examples are Sets and Maps, which
are implemented by self-balancing binary trees. The representation in
memory of such data structures is dependent on the order in which
values are inserted into them, however their equality is based entirely
upon their contents. Using Sets of ints as an example
`(add 1 (add 2 empty))` may be structurally different from
`(add 2 (add 1 empty))` even though these values are considered to be
equal.

Thus it is incorrect to use `Poly.hash` to compute the hash of a Set or
Map.

This change updates the hash computation for the `Pkg_rules.DB.t` type
to avoid the use of `Poly.hash` as well as several other types which
comprise it.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>