test: fix exec-watch-server test and demonstrate issue
The timeout set in this test was way too high. I set it back to a
sensible level so that when it fails it doesn't stall. I've also added a
timeout to the exec-watch tests in general, since there shouldn't be any
slow tests.

The exec-watch-server test has been reenabled and the issue in
https://github.com/ocaml/dune/issues/12323 is demonstrated.

Signed-off-by: Ali Caglayan <alizter@gmail.com>