Remove most backtracking from lockdir parser
To parse both portable and non-portable lockdirs with mostly the same
code, the parser would attempt to parse fields with the non-portable
lockdir syntax and fall back to the portable syntax should parsing fail.
This led to the generation of parse errors when the first attempt at
parsing a field would fail, and even though these errors are discarded,
generating them took a noticeable amount of time.

This changes the parsing logic to only attempt to parse fields in a
single syntax, determined by the presence of the `solved_for_platforms`
field in the lockdir metadata file. This field is present precisely when
a lockdir is portable.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>