Don't use Digest.generic for toolchain hash
The Digest.generic function marshals a value to a string and then hashes
the string. For types which contain sets and maps, two values
constructed in different orders but which are considered equal may have
different in-memory representations and so marshal to different strings
and hash to different hashes.

Compiler toolchain directories created by dune have the package hash
appended to their names. Using Digest.generic meant that sometimes Dune
would create a new toolchain directory when re-using an existing one
would suffice.

The solution is to implement a hash function for packages which does not
depend on the memory representation of packages. To simplify
implementing hash functions for custom types with blake3, this change
also introduces the `Digest.Feed` module with some helper functions for
composing hash functions for different types.

Note that this change alone will not fix
https://github.com/ocaml/dune/issues/12390.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>