pkg: further improve performance of null builds (#12248)
A previous pr that replaced uses of polymorphic hash on sets also slowed
down dune somewhat, because Poly.hash on a large set hashes a few values and
gives up, whereas the hand-written hash function hashes the whole value.

Rather than replicating the behavior of Poly.hash, the simplest thing to do
is to not hash such big values.

In the test repo from #12248, a null `dune build` goes from 1.8s to 1s with
this change.

Signed-off-by: Valentin Gatien-Baron <valentin.gatienbaron@gmail.com>