Add a `dune tools env` command to add dev tools to PATH
The expected use-case is for the output of this command to be eval'd in
the user's shell. In order for a terminal-based editor to run the
ocamllsp dev tool, `eval $(dune tools env)` must be run in that terminal
before launching the editor.

Fixes https://github.com/ocaml/dune/issues/10964

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>