Enable package management for more tests
Going forward the signal to use package management is `(pkg enabled)` so
the tests that currently rely on it always enabled/autodetected need to
opt in.

Fortunately, on `main` the meaning of "unset" and `enabled` are
identical in cases where package management is used, so it can be just
enabled.

Signed-off-by: Marek Kubica <marek@tarides.com>