Reuse dependencies between project and tools (#12526)
This allows the dependencies of the project to be shared with dev tools,
preventing unnecessarily rebuilding dependencies while building a tool
if an identical version of the package has already been built as a
dependency of the project or vice versa.

This requires storing build artifacts of dependencies of dev tools and
the project (when built in the "default" context) in a single directory,
since their relative path within the _build directory is part of the key
used to cache build artifacts. Previously the build artifacts of a
package were placed in a directory named after that package, however
since dev tools and the project may depend on different versions of the
same package, directories needed to be renamed to avoid collisions. Even
if the project and dev tools rely on the same version of some package,
it may still need to be rebuilt on occasion due to the presence/absence
of depopts, or due to manual modification to lockfiles. For this reason,
the directories containing package build artifacts contain in their
name, a hash of the contents of the lockfiles in the package's transitive
dependency closure.

This change introduces the concept of a "package digest" which is a
unique directory name for a package comprising its name, version, and
dependency closure hash. The DB type used to store metadata while
building packages has been updated to be keyed based on package digests
(rather than package names) and to store package metadata for dev tools
and the project's default build context in the same table to allow
dependencies to be shared.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>