ci: update for ocaml 5.4 release (#12552)
Update the CI workflows to track the current major version trunk and
use the latest release in the build matrix, when tests are not run.

We cannot currently udate to 5.4 for the matrix used to run tests, but
some of the test dependencies don't support it.

As per https://github.com/ocaml/opam-repository/pull/28683

Signed-off-by: Shon Feder <shon.feder@gmail.com>