nix: devShell for oxcaml
We add 3 new dev shells for OxCaml:

- bootstrap-ox: For testing the bootstrap process with OxCaml
- ox-minimal: For building dune with OxCaml and running the OxCaml test-cases
- ox: For attempting to build a dev shell with OxCaml

In order to do this, we need to fetch the latest patches from the OxCaml
opam repo, however not all of them can be applied to the package
versions found in nixpkgs. Because of this we cannot build a full
devShell, but we can get pretty close. The File nix/ox-patches.nix deals
with the fetching and application of the patches.

Signed-off-by: Ali Caglayan <alizter@gmail.com>