fix(pkg): make dune pin itelf
Automatically make dune pin itself. We're going to assume that the dune
we're using right now is the one that will be used to run the build
plan.

Perhaps this behavior isn't always deired. We could imagine wanting to
produce build plans for older or newer dunes. For now, this is a good
default though.

Addresses #12381

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>