pkg: relock ocamlformat when version changes
The .ocamlformat file can contain the version of ocamlformat which a
project uses for autoformatting. Previously this file was used to
determine which version of ocamlformat to install initially, but changes
to the version specified in .ocamlformat did not cause dune to relock
the tool with the new version, leading to a mismatch between the
specified and installed versions of ocamlformat within a project.

This change allows dune to detect when the specified and locked versions
of ocamlformat differ, and relocks (and re-installs) ocamlformat in this
case.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>