Dune Pins itself (#12583)
* Reproduce issue in #12381

The version of dune inside the opam repo is overriding the version of
dune that is being used. That seems wrong. At least most of the time.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>

* Remove [Resolved_package.set_url]

Fuse it with [git_repo] since that is the only place where it's needed

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>

* fix(pkg): make dune pin itelf

Automatically make dune pin itself. We're going to assume that the dune
we're using right now is the one that will be used to run the build
plan.

Perhaps this behavior isn't always deired. We could imagine wanting to
produce build plans for older or newer dunes. For now, this is a good
default though.

Addresses #12381

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>