fix: promotion over non-existent file
When we use "git diff" as the git tool, we have some logic to resolve
symlinks since "git diff" doesn't handle them. This logic doesn't take
into account the case where the target of a symlink doesn't exist, which
happens when the file you are promoting doesn't exist yet.

This can easily be fixed by taking into account this case.

Signed-off-by: Ali Caglayan <alizter@gmail.com>