pkg: print portable solutions grouped by platform
The message that prints after solving lists each package in the
solution. When a solution is portable, it's possible (though hopefully
rare) that the version of a package included in a solution will be
different depending on the platform. This led to multiple versions of
the same package appearing in the message for non-obvious reasons, which
some found confusing.

This commit changes the message when portable lockdirs are enabled, such
that the packages common to all platforms are printed together, and each
set of platform-specific dependencies is printetd separately from the
common dependencies.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>