Fix package validation when portable lockdirs is enabled
The `dune pkg validate-lockdir` command checks that the lockdir contains
the transitive dependency closure of all local packages. When run with
portable lockdirs enabled, some packages were omitted from the
dependency closure because the wrong solver environment was being
treated as the current platform. This change fixes the issue by using
the correct solver environment to represent the current platform, and
does a minor refactor so the validation logic isn't juggling two
different solver environments (the second environment is the one stored
in the lockdir capturing the opam solver variables used to generate the
solution).

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>