Deduplicate solver errors when generating portable lockdirs
When generating portable lockdirs the solver runs multiple times. Dune
collects any errors from the solver and prints them to the terminal in
the event that no solution was found on any platforms, and logs them
printing a warning if no solution was found on some but not all
platforms. This can create a situation where the same error is printed
multiple times, since many solver errors are platform agnostic.

This change groups solver errors by the contents of their message when
generating portable lockdirs. Each solver error is printed a single
time, along with the list of platforms on which it occurred.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>