Fix dependencies of utop dev tool
This fixes a bug in computing dependencies of a dev tool package when
using dune package management. It only affected the utop dev tool. This
change also adds an integration test for the utop dev tool. The test is
a github action workflow rather than a cram test so that the real utop
package can be tested, since the utop rules have some logic which is
specific to the utop package and thus is non-trivial to accurately test
with a mock utop package as we do with other dev tool tests.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>