Network caps to track network access
Adds a [Network_cap.t] type which must be passed to the low-level
functions which require network access, and should only be created at
the "edge", either in a command or a rule. This will make it harder to
accidentally add a requirement on network access to a code path that
previously did not require network access, which will become important
for maintaining an offline mode for Dune where it avoids updating
package repositories and downloading files where possible.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>