pkg: don't recreate Pkg_rules.DB values
This table is only created and used twice. We add an Id.t field so that
it can easily be compared later when it's values are memoized in the
of_ctx call.

The old hash and equality functions were expensive an unnecessary.

We also memoize the of_dev_tool caller

Before:
```
[ocaml-cohttp]$ hyperfine "dune build"
Benchmark 1: dune build
  Time (mean ± σ):      1.851 s ±  0.039 s    [User: 1.561 s, System: 0.283 s]
  Range (min … max):    1.814 s …  1.943 s    10 runs
```

After:
```
[ocaml-cohttp]$ hyperfine "dune build"
Benchmark 1: dune build
  Time (mean ± σ):      1.086 s ±  0.016 s    [User: 0.803 s, System: 0.279 s]
  Range (min … max):    1.052 s …  1.108 s    10 runs
```

Signed-off-by: Ali Caglayan <alizter@gmail.com>