Dune_trace: emit based on category (#12940)
Introduce a type for all the categories we emit.

We check if a category is enabled before trying to generate & emit a
particular event.

For now, the only way to control categories is via an environment
variable. This isn't meant to be user facing, so this isn't documented
yet. All events that users will care about will be enabled by default.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>