Don't hardcode compiler packages for dev tools (#12834)
This adds as dependencies more compiler constraints to dev tools if it
needs to use the same compiler.

I was discussing with @Alizter last week that when you try to install
`ocamllsp` in a project built with OxCaml, it doesn't automatically
pickup the OxCaml version of `ocaml-lsp-server` and that one has to
explicitly specify this in `dune-workspace`. Digging further, I found
that the package `ocaml` is only considered as a compiler package,
whereas OxCaml and other compiler branches exist as `ocaml-variants`. To
this end, it seemed like it would make sense to add them as compiler
dependencies.

This PR is not in the best shape, but I'm opening it early to ask if
this is something we should be doing while locking dev tools. If this is
worth adding, I'll work on refactoring this a bit, add a cram test, and
make it ready for review.

I've run some basic tests with this branch and it does pickup the right
dev tools on OxCaml projects.

---------

Signed-off-by: Sudha Parimala <sudharg247@gmail.com>
Signed-off-by: Marek Kubica <marek@tarides.com>
Co-authored-by: Marek Kubica <marek@tarides.com>