Portable `sed`-like functionality (#13008)
Using `sed` is somewhat temperamental: despite POSIX, the CLI is notably
different between BSD and GNU systems making it tedious to use as other
systems don't support the options, the regular expression support is
dependant on arguments, `sed` scripts are fairly difficult to read (with
`/` and `#` separators for paths instead of just command line arguments)
and mostly only used to do simple replacements (so often just overkill).

This PR adds a few new commands to `dune_cmd` to implement `sed`-like
functionality but in a way that is consistent between GNU/Linux, BSDs,
macOS and potentially Windows. It supports in-place replacement, since
this is what we usually wants in our cram tests as well as interactions
via stdio.

Other `sed` functionality could be added, it's implemented in somewhat
extensible way by parsing into a `Sed.t` action and evaluating that.

I've changed some of the helpers from the `dune pkg` tests to use it.

Signed-off-by: Marek Kubica <marek@tarides.com>