Structured Logging (#13015)
Instead of maintaining a separate _build/log, we fold it into the trace
and produce it by default. The trace already has a lot more valuable
information than the log, so the combined information provides a better
debugging experienced. Log messages also contain structured payloads,
which makes them useful with automation/scripting. This is demonstrated
in the changes in the tests.

Fixes #12904 except for the last point. That will come in a subsequent
PR.

---------

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>