Fix test suite outside of macos (#13073)
`enabled_if` on cram tests applies the `enabled_if` on the tests that it
matches rather than on the `cram` statement itself.

It's a shame that we don't have a good way to accomplish what you've
tried here. The best I can think of is some sort `:include` hackery in
the list of deps that generates the dependency conditionally.

cc @anmonteiro

---------

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>