Nested dune call fix (executable resolution part) (#12902)
#10019 mentions that Dune should run itself when encountering package
rules, which is quite sensible to do. This PR fixes part of the problem
- the `run` command resolution. When encountering a command dune
attempts to look up the programs path. But if the command is `dune`, we
can immediately resolve this to the current executable (was we are Dune,
even if the command is not called `dune` but `dune.exe` or `main.exe` or
something else). This prevents capture of an accidental dune executable
(as exhibited by the repro fix).

This PR thus solves the first case of #10020, where `dune` is called
directly. It does not fix the second part (dune not part of the path),
which requires a different fix that I intend to submit as a separate PR
(as the fix is somewhat independent from this one).

---------

Signed-off-by: Marek Kubica <marek@tarides.com>