Simplify test checking the pinning behavior (#13106)
I've been investigating a report by @samoht and looked whether we have a
test for the behavior. We have a test but it documents the opposite
behavior of what @samoht wants.

This PR cleans out the test so it doesn't require `make` and thus it
does not need to deal with normalizing `make` output.

Signed-off-by: Marek Kubica <marek@tarides.com>