Add test showing the current behavior of no-prefix, `file` and `git` pins (#13127)
This test demonstrates the current behavior of locking different types
of URLs. It is part of #13110 which should change the behavior going
forward.

I chose not to inspect lock files to show the contents but rather just
document the behavior, as that saves me from dealing with git commit
hashes.

Signed-off-by: Marek Kubica <marek@tarides.com>